ref(frontend): Rename Menu to MainMenu
