+++ frontend/src/App.vue
@@ -1,15 +1,15 @@
 <template>
-    <Menu />
+    <MainMenu />
     <div class="max-w-7xl mx-auto p-4 md:p-8">
         <RouterView />
     </div>
 </template>
 
 <script>
-import Menu from '@/components/Menu.vue';
+import MainMenu from '@/components/MainMenu.vue';
 
 export default {
     name: "App",
-    components: { Menu },
+    components: { MainMenu },
 };
 </script>
