+++ frontend/src/components/MainMenu.vue
